requirements.txt
Django==1.9
djangorestframework==3.3.1
googlemaps==2.4.2
requests==2.6.0
wheel==0.24.0
